last update

+++ quick-location-ws/firebase-integration/src/main/java/com/quick/location/model/firebase/TopTenFirebase.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 public class TopTenFirebase {
     
     String name;
-    long count;
-    long porcent;
+    double count;
+    double porcent;
 
 }
